Move `getCanvasBlob`
