+++ packages/pdf/src/index.ts
@@ -1,9 +1,37 @@
 import { GlobalWorkerOptions, getDocument } from 'pdfjs-dist';
 
-import { getCanvasBlob, getExtensionFromMimeType } from '@mexp/media-utils';
+import { getExtensionFromMimeType } from '@mexp/media-utils';
 
 // Setting worker path to worker bundle.
 GlobalWorkerOptions.workerSrc = PDFJS_CDN_URL;
+
+/**
+ * Returns a Blob object from a canvas element.
+ *
+ * Simple, promise-ified version of `HTMLCanvasElement.toBlob()`.
+ *
+ * @see https://developer.mozilla.org/en-US/docs/Web/API/HTMLCanvasElement/toBlob
+ *
+ * @param canvas  Canvas element
+ * @param type    Desired image format.
+ * @param quality Desired image quality.
+ */
+function getCanvasBlob(
+	canvas: HTMLCanvasElement,
+	type = 'image/jpeg',
+	quality = 0.82
+): Promise< Blob > {
+	return new Promise< Blob >( ( resolve, reject ) => {
+		canvas.toBlob(
+			( blob ) =>
+				blob
+					? resolve( blob )
+					: reject( new Error( 'Could not get canvas blob' ) ),
+			type,
+			quality
+		);
+	} );
+}
 
 /**
  * Returns an image file for a given PDF.
